+++ packages/types/src/options.ts
@@ -1,4 +1,4 @@
-import type { ReportData } from '@null/monitor-types'
+import type { ReportData } from '@zjy/monitor-types'
 import type { BreadcrumbData } from './breadcrumb'
 
 export interface InitOptions {