+++ packages/web/src/index.ts
@@ -1,4 +1,4 @@
-import { init, log, SDK_NAME, SDK_VERSION } from '@null/monitor-browser'
-import { MonitorVue } from '@null/monitor-vue'
+import { init, log, SDK_NAME, SDK_VERSION } from '@zjy/monitor-browser'
+import { MonitorVue } from '@zjy/monitor-vue'
 
 export { init, log, MonitorVue, SDK_NAME, SDK_VERSION }