+++ packages/web/CHANGELOG.md
@@ -9,5 +9,5 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @null/monitor-browser@0.1.0
-  - @null/monitor-vue@0.1.0
+  - @zjy/monitor-browser@0.1.0
+  - @zjy/monitor-vue@0.1.0